requirements.txt
Flask
pytesseract
Pillow
openai
python-dotenv
flask_cors
httplib2
six
python-taiga